requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
frozenlist==1.7.0
greenlet==3.2.3
gunicorn==21.2.0
idna==3.10
multidict==6.6.3
packaging==25.0
pillow==11.3.0
propcache==0.3.2
PyJWT==2.10.1
requests==2.32.4
SQLAlchemy==2.0.41
sqlparse==0.5.3
twilio==9.6.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.9.0
yarl==1.20.1
